Unified, self-synchronizing tx subscriber.
xtdb.tx.subscribe now has a single subscriber implementation, which
supports both asynchronous notifications and polling, as needed. In
addition, it does not depend on a counting semaphore, so race conditions
such as database reconnects won't cause anything to get out of sync.
Notifying the subscriber handler of a transaction is always safe, even
if it's redundant or out of order.